# docs/swagger.yaml
basePath: /panel/api
definitions:
...
  entity.AllSetting:
    properties:
      datepicker:
        description: Date picker format
        type: string
      expireDiff:
        description: Expiration warning threshold in days
        type: integer
      externalTrafficInformEnable:
        description: Enable external traffic reporting
        type: boolean
      externalTrafficInformURI:
        description: URI for external traffic reporting
        type: string
      ldapAutoCreate:
        type: boolean
      ldapAutoDelete:
        type: boolean
      ldapBaseDN:
        type: string
      ldapBindDN:
        type: string
      ldapDefaultExpiryDays:
        type: integer
      ldapDefaultLimitIP:
        type: integer
      ldapDefaultTotalGB:
        type: integer
      ldapEnable:
        description: LDAP settings
        type: boolean
      ldapFlagField:
        description: Generic flag configuration
        type: string
      ldapHost:
        type: string
      ldapInboundTags:
        type: string
      ldapInvertFlag:
        type: boolean
      ldapPassword:
        type: string
      ldapPort:
        type: integer
      ldapSyncCron:
        type: string
      ldapTruthyValues:
        type: string
      ldapUseTLS:
        type: boolean
      ldapUserAttr:
        description: e.g., mail or uid
        type: string
      ldapUserFilter:
        type: string
      ldapVlessField:
        type: string
      pageSize:
        description: UI settings
        type: integer
      remarkModel:
        description: Remark model pattern for inbounds
        type: string
      sessionMaxAge:
        description: Session maximum age in minutes
        type: integer
      subCertFile:
        description: SSL certificate file for subscription server
        type: string
      subDomain:
        description: Domain for subscription server validation
        type: string
      subEnable:
        description: Subscription server settings
        type: boolean
      subEncrypt:
        description: Encrypt subscription responses
        type: boolean
      subJsonEnable:
        description: Enable JSON subscription endpoint
        type: boolean
      subJsonFragment:
        description: JSON subscription fragment configuration
        type: string
      subJsonMux:
        description: JSON subscription mux configuration
        type: string
      subJsonNoises:
        description: JSON subscription noise configuration
        type: string
      subJsonPath:
        description: Path for JSON subscription endpoint
        type: string
      subJsonRules:
        type: string
      subJsonURI:
        description: JSON subscription server URI
        type: string
      subKeyFile:
        description: SSL private key file for subscription server
        type: string
      subListen:
        description: Subscription server listen IP
        type: string
      subPath:
        description: Base path for subscription URLs
        type: string
      subPort:
        description: Subscription server port
        type: integer
      subShowInfo:
        description: Show client information in subscriptions
        type: boolean
      subTitle:
        description: Subscription title
        type: string
      subURI:
        description: Subscription server URI
        type: string
      subUpdates:
        description: Subscription update interval in minutes
        type: integer
      swaggerEnable:
        description: Enable Swagger API documentation
        type: boolean
      tgBotAPIServer:
        description: Custom API server for Telegram bot
        type: string
      tgBotBackup:
        description: Enable database backup via Telegram
        type: boolean
      tgBotChatId:
        description: Telegram chat ID for notifications
        type: string
      tgBotEnable:
        description: Telegram bot settings
        type: boolean
      tgBotLoginNotify:
        description: Send login notifications
        type: boolean
      tgBotProxy:
        description: Proxy URL for Telegram bot
        type: string
      tgBotToken:
        description: Telegram bot token
        type: string
      tgCpu:
        description: CPU usage threshold for alerts
        type: integer
      tgLang:
        description: Telegram bot language
        type: string
      tgRunTime:
        description: Cron schedule for Telegram notifications
        type: string
      timeLocation:
        description: Security settings
        type: string
      trafficDiff:
        description: Traffic warning threshold percentage
        type: integer
      twoFactorEnable:
        description: Enable two-factor authentication
        type: boolean
      twoFactorToken:
        description: Two-factor authentication token
        type: string
      webBasePath:
        description: Base path for web panel URLs
        type: string
      webCertFile:
        description: Path to SSL certificate file for web server
        type: string
      webDomain:
        description: Web server domain for domain validation
        type: string
      webKeyFile:
        description: Path to SSL private key file for web server
        type: string
      webListen:
        description: Web server settings
        type: string
      webPort:
        description: Web server port number
        type: integer
    type: object
  entity.Msg:
    properties:
      msg:
        description: Response message text
        type: string
      obj:
        description: Optional data object
      success:
        description: Indicates if the operation was successful
        type: boolean
    type: object
...
host: localhost:2053
info:
  contact:
    name: API Support
    url: https://github.com/ByteProvider/3x-ui/issues
  description: API documentation for 3x-ui panel - Xray management system
  license:
    name: GPL-3.0
    url: https://github.com/ByteProvider/3x-ui/blob/main/LICENSE
  termsOfService: https://github.com/ByteProvider/3x-ui
  title: 3x-ui Panel API
  version: "2.0"
paths:
...
  /setting/getDefaultJsonConfig:
    get:
      consumes:
      - application/json
      description: Retrieve the default Xray configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get default Xray config
      tags:
      - settings
...
  /setting/update:
    post:
      consumes:
      - application/json
      description: Update all settings with the provided data
      parameters:
      - description: All settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/entity.AllSetting'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Update settings
      tags:
      - settings
...
  /xray/:
    post:
      consumes:
      - application/json
      description: Retrieve the Xray configuration template and inbound tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get Xray settings
      tags:
      - xray
  /xray/getDefaultJsonConfig:
    get:
      consumes:
      - application/json
      description: Retrieve the default Xray configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get default Xray config
      tags:
      - xray
  /xray/getOutboundsTraffic:
    get:
      consumes:
      - application/json
      description: Retrieve the traffic statistics for outbounds
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get outbounds traffic
      tags:
      - xray
  /xray/getXrayResult:
    get:
      consumes:
      - application/json
      description: Retrieve the current Xray service result
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get Xray result
      tags:
      - xray
  /xray/resetOutboundsTraffic:
    post:
      consumes:
      - application/json
      description: Reset the traffic statistics for the specified outbound tag
      parameters:
      - description: Outbound tag
        in: formData
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Reset outbound traffic
      tags:
      - xray
  /xray/update:
    post:
      consumes:
      - application/json
      description: Update the Xray configuration settings
      parameters:
      - description: Xray settings JSON
        in: formData
        name: xraySetting
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Update Xray settings
      tags:
      - xray
  /xray/warp/{action}:
    post:
      consumes:
      - application/json
      description: Handle Warp-related operations (data, del, config, reg, license)
      parameters:
      - description: Warp action (data, del, config, reg, license)
        in: path
        name: action
        required: true
        type: string
      - description: Private key (for reg action)
        in: formData
        name: privateKey
        type: string
      - description: Public key (for reg action)
        in: formData
        name: publicKey
        type: string
      - description: License (for license action)
        in: formData
        name: license
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Handle Warp operations
      tags:
      - xray
securityDefinitions:
  ApiKeyAuth:
    description: API Key authentication. Use X-API-Key header or Authorization Bearer
      token
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"